pyproject.toml
[tool.poetry]
name = "store-management-system"
version = "0.1.0"
description = ""
authors = ["shubh220922 <shubh22.srivastava@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.0"
python-dotenv = "^1.0.0"
mysql-connector-python = "^8.1.0"
sqlalchemy = "^2.0.21"
pymysql = "^1.1.0"
bcrypt = "^4.0.1"


[tool.poetry.group.dev.dependencies]
setuptools = "^68.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
